Rework docs & examples
