+++ docs_src/async_sql_databases/tutorial001.py
@@ -1,15 +1,18 @@
-from typing import List
+from typing import AsyncGenerator, List
 
 import databases
 import sqlalchemy
-from fastapi import FastAPI
-from pydantic import BaseModel
+from fastapi import Depends, FastAPI
+from pydantic import BaseModel, BaseSettings
 
-# SQLAlchemy specific code, as with any other app
-DATABASE_URL = "sqlite:///./test.db"
-# DATABASE_URL = "postgresql://user:password@postgresserver/db"
 
-database = databases.Database(DATABASE_URL)
+class Settings(BaseSettings):
+    db_url: str
+
+
+def get_config() -> Settings:
+    return Settings()
+
 
 metadata = sqlalchemy.MetaData()
 
@@ -22,10 +25,18 @@ notes = sqlalchemy.Table(
 )
 
 
-engine = sqlalchemy.create_engine(
-    DATABASE_URL, connect_args={"check_same_thread": False}
-)
-metadata.create_all(engine)
+def setup_schema(config: Settings = Depends(get_config)) -> None:
+    engine = sqlalchemy.create_engine(
+        config.db_url, connect_args={"check_same_thread": False}
+    )
+    metadata.create_all(engine)
+
+
+async def get_db(
+    config: Settings = Depends(get_config), schema: None = Depends(setup_schema)
+) -> AsyncGenerator[databases.Database, None]:
+    async with databases.Database(config.db_url) as db:
+        yield db
 
 
 class NoteIn(BaseModel):
@@ -39,27 +50,23 @@ class Note(BaseModel):
     completed: bool
 
 
-app = FastAPI()
+async def startup(
+    db: databases.Database = Depends(get_db, use_cache="app", lifetime="app")
+):
+    db.execute("SELECT 1")  # checks that we are properly connected to the database
 
 
-@app.on_event("startup")
-async def startup():
-    await database.connect()
-
-
-@app.on_event("shutdown")
-async def shutdown():
-    await database.disconnect()
+app = FastAPI(on_startup=[startup])
 
 
 @app.get("/notes/", response_model=List[Note])
-async def read_notes():
+async def read_notes(db: databases.Database = Depends(get_db)):
     query = notes.select()
-    return await database.fetch_all(query)
+    return await db.fetch_all(query)
 
 
 @app.post("/notes/", response_model=Note)
-async def create_note(note: NoteIn):
+async def create_note(note: NoteIn, db: databases.Database = Depends(get_db)):
     query = notes.insert().values(text=note.text, completed=note.completed)
-    last_record_id = await database.execute(query)
+    last_record_id = await db.execute(query)
     return {**note.dict(), "id": last_record_id}
