--- docs_src/dependencies/app_lifespan_db/config.py
@@ -1,9 +1,0 @@
-from pydantic import BaseSettings
-
-
-class Settings(BaseSettings):
-    db_url: str = "postgresql://user:password@postgresserver/db"
-
-
-def get_config() -> Settings:
-    return Settings()
