--- docs_src/dependencies/app_lifespan_db/models.py
@@ -1,12 +1,0 @@
-from pydantic import BaseModel
-
-
-class NoteIn(BaseModel):
-    text: str
-    completed: bool
-
-
-class Note(BaseModel):
-    id: int
-    text: str
-    completed: bool
