--- docs_src/dependencies/app_lifespan_db/schemas.py
@@ -1,12 +1,0 @@
-import sqlalchemy
-
-metadata = sqlalchemy.MetaData()
-
-
-notes = sqlalchemy.Table(
-    "notes",
-    metadata,
-    sqlalchemy.Column("id", sqlalchemy.Integer, primary_key=True),
-    sqlalchemy.Column("text", sqlalchemy.String),
-    sqlalchemy.Column("completed", sqlalchemy.Boolean),
-)
