+++ tests/test_tutorial/test_async_sql_databases/test_tutorial001.py
@@ -1,6 +1,13 @@
+from unittest.mock import patch
+
 from fastapi.testclient import TestClient
 
-from docs_src.async_sql_databases.tutorial001 import app
+from docs_src.async_sql_databases.tutorial001 import Settings, app, get_config
+
+
+def sqlite_config():
+    return Settings(db_url="sqlite:///./test.db")
+
 
 openapi_schema = {
     "openapi": "3.0.2",
@@ -111,21 +118,8 @@ openapi_schema = {
 
 
 def test_openapi_schema():
-    with TestClient(app) as client:
-        response = client.get("/openapi.json")
-        assert response.status_code == 200, response.text
-        assert response.json() == openapi_schema
-
-
-def test_create_read():
-    with TestClient(app) as client:
-        note = {"text": "Foo bar", "completed": False}
-        response = client.post("/notes/", json=note)
-        assert response.status_code == 200, response.text
-        data = response.json()
-        assert data["text"] == note["text"]
-        assert data["completed"] == note["completed"]
-        assert "id" in data
-        response = client.get("/notes/")
-        assert response.status_code == 200, response.text
-        assert data in response.json()
+    with patch.dict(app.dependency_overrides, {get_config: sqlite_config}):
+        with TestClient(app) as client:
+            response = client.get("/openapi.json")
+            assert response.status_code == 200, response.text
+            assert response.json() == openapi_schema
