+++ tests/test_tutorial/test_dependencies/test_tutorial014.py
@@ -1,0 +1,8 @@
+from fastapi.testclient import TestClient
+
+from docs_src.dependencies.tutorial014 import app
+
+
+def test_tutorial_13():
+    with TestClient(app):
+        ...  # pragma: no cover
